+++ packages/mcp2doql/src/mcp2doql/__init__.py
@@ -1,0 +1,6 @@
+from .scanner import scan_python_mcp, discover_mcp_tools
+
+__all__ = [
+    "scan_python_mcp",
+    "discover_mcp_tools",
+]